SDK-600: Add further validation, make Yoti host constants public
